Defer command additions for sub-commands until their parent has been registered.
If a sub-command is being added, this addition will be deferred when its parent is not yet registered. This is done by hooking the actual addition up to the `after_add_command:$parent` hook.

The list of deferred commands is stored to keep track of commands for which the parent hook is never called. This is the case for commands that provide a nested sub-command only, like `wp network meta`.

The bootstrap process has an additional step `RegisterDeferredCommands` that goes through the list of outstanding commands and adds them, as there's no opportunity for the parent being registered at this point.

Fixes #4084