Ensure the autoloader for the Requests library is registered
When a command uses `Utils\http_request()` when WordPress Core has not been loaded yet, we need to ensure that the autoloader for the Requests library is actually registered before doing the request.

Fixes #5295

This PR also does a bit of clean-up to make the dependencies of the `Utils` class a bit more obvious.